refactor(Reader,ReaderValidator): 입력값 검증 클래스 분리
기존 내부 클래스로 정의되었던 입력값 검증 클래스를 분리하여 정의
