+++ src/main/java/bridge/view/console/ReaderValidator.java
@@ -1,0 +1,16 @@
+package bridge.view.console;
+
+import bridge.Validator;
+
+public class ReaderValidator extends Validator {
+    public static void validateNumber(String value) {
+        validateNotNull(value);
+        validateNotEmpty(value);
+        validateInteger(value);
+    }
+
+    public static void validateString(String value) {
+        validateNotNull(value);
+        validateNotEmpty(value);
+    }
+}
